Renames.
Packages properly named.
Entrypoint properly named.
